% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_flows.R
\name{approx_flows}
\alias{approx_flows}
\title{approx_flows}
\usage{
approx_flows(time, in_E, out_E, in_H, out_H, wb_E, wb_H)
}
\arguments{
\item{time}{numeric}

\item{in_E}{numeric}

\item{out_E}{numeric}

\item{in_H}{numeric}

\item{out_H}{numeric}

\item{wb_E}{numeric}

\item{wb_H}{numeric}
}
\description{
approx_flows
}
