% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge_summerflood.R
\docType{data}
\name{discharge_summerflood}
\alias{discharge_summerflood}
\title{Inflow and Outflow Data}
\format{
Data frame with the following columns:
\describe{
  \item{time}{time in days}
  \item{inflow}{water inflow}
  \item{outflow}{managed outflow, e.g. for drinking water supply}
  \item{outflow_wb}{outflow to the river bed}
}
}
\source{
Lukas Gunzelmann, derived from real measurements
}
\description{
Example data sets with inflow and outflow of a reservoir (in Mio m3/d)
of a theoretical year with a high flood in summer.
}
\examples{

data("discharge_summerflood")
}
\keyword{data}
