% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_data.R, R/discharge_wet.R
\docType{data}
\name{discharge_wet}
\alias{discharge_wet}
\title{Data set discharge_wet}
\format{
Data frame with the following columns:
\describe{
  \item{time}{time in days}
  \item{inflow}{water inflow}
  \item{outflow}{managed outflow, e.g. for drinking water supply}
  \item{outflow_wb}{outflow to the river bed}
}
}
\source{
Lukas Gunzelmann, derived from real measurements
}
\description{
Example Hydrological Data of d Wet Year

Example data sets with inflow and outflow of a reservoir (in Mio m3/d)
of a normal, a wet and a dry year.
}
\examples{

data("discharge_wet")

}
\keyword{data}
