% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flower.R
\name{flower}
\alias{flower}
\title{Launch outfloweR shiny App}
\usage{
flower(...)
}
\arguments{
\item{...}{optional parameters passed to the app (not yet implemented)}
}
\value{
This function has no return value.
}
\description{
Launch a shiny app to visualize mass balance of reservoirs.
}
\details{
The app is still experimental.
}
\examples{

\dontrun{
flower()
}

}
