% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outflower-package.R
\docType{package}
\name{growthrates-package}
\alias{growthrates-package}
\alias{outfloweR}
\alias{outfloweR-package}
\title{\packageTitle{outfloweR}}
\description{
\packageDescription{outfloweR}

The package aims to demonstrate reservoir management options by abstracting
water either from the bottom outlet or the surface layer (the epilimnion).
It uses a simple mass-balance approach and provides a shiny-app.
}
\references{
ATT-Schriftenreihe Bd. 7: Integrale Bewirtschaftung von Trinkwassertalsperren
gemäß DIN 19700. Siegburg, 2009.
#'
}
\author{
Thomas Petzoldt, Lukas Gunzelmann
}
\keyword{package}
