% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{scenario}
\alias{scenario}
\title{Run scenarions of a reservoir mass balance model}
\usage{
scenario(
  option = c("standard", "cold_inflow", "wb_top"),
  vol,
  t_strat,
  t_autumn,
  inflow,
  outflow,
  outflow_wb
)
}
\arguments{
\item{option}{character, specifying the water inflow and water abstraction regime}

\item{vol}{numeric value, volume of the reservoir (in Mio m3)}

\item{t_strat}{numeric value, day of stratification onset}

\item{t_autumn}{numeri value,  end of stratification}

\item{inflow}{numeric vector, inflow in Mio m3/d}

\item{outflow}{numeric vector, outflow in Mio m3/d}

\item{outflow_wb}{numeric vector, bottom outlet in Mio m3/d}
}
\description{
Run scenarions of a reservoir mass balance model
}
