% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratification_event.R
\name{stratification_event}
\alias{stratification_event}
\title{stratification_event}
\usage{
stratification_event(time, y, parms, inflows)
}
\arguments{
\item{time}{numeric}

\item{y}{numeric}

\item{parms}{numeric}

\item{inflows}{list}
}
\description{
stratification_event
}
