% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_model.R
\name{volume_model}
\alias{volume_model}
\title{volume_model}
\usage{
volume_model(time, y, parms, inflows)
}
\arguments{
\item{time}{numeric}

\item{y}{numeric}

\item{parms}{numeric}

\item{inflows}{list}
}
\description{
volume_model
}
