% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volumes.R
\name{plot_volumes}
\alias{plot_volumes}
\title{plot_volumes}
\usage{
plot_volumes(out, ylab = "Volumen (Mio m3)")
}
\arguments{
\item{out}{output from `ode` (deSolve Object)}

\item{ylab}{label at the y axis}
}
\description{
plot_volumes
}
